requirements.txt
# 投资日报
python-dotenv>=1.0
feedparser>=6.0
yfinance>=0.2.40
requests>=2.28
beautifulsoup4>=4.12
